% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab_server}
\alias{admtab_server}
\title{Serverdel av modul for Administrative tabeller-fane i
NNRR sin shiny-app på Rapporteket}
\usage{
admtab_server(id, RegData, userRole, hvd_session)
}
\value{
Modulfunksjoner til Administrative tabeller
}
\description{
Kun til bruk i Shiny
}
