% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab_ui}
\alias{admtab_ui}
\title{UI-modul for Administrative tabeller-fane i NNRR sin
shiny-app på Rapporteket}
\usage{
admtab_ui(id)
}
\value{
Modulfunksjoner til Administrative tabeller
}
\description{
Kun til bruk i Shiny
}
