% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_Server}
\alias{datadump_Server}
\title{Modul for serverdel av datadump i NNRR sin shiny-app på Rapporteket}
\usage{
datadump_Server(id, RegData, hvd_session, user)
}
\description{
Kun til bruk i Shiny
}
