% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_UI}
\alias{fordelingsfig_UI}
\title{UI-modul for fordelingsfigurer i NNRR sin shiny-app på Rapporteket}
\usage{
fordelingsfig_UI(id, BrValg)
}
\description{
Kun til bruk i Shiny
}
