% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfigServer}
\alias{fordelingsfigServer}
\title{Server-modul for fordelingsfigurer i NNRR sin shiny-app på Rapporteket}
\usage{
fordelingsfigServer(id, reshID, RegData, userRole, hvd_session)
}
\description{
Kun til bruk i Shiny
}
