% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_indikatorfig.R
\name{indikatorfigServer}
\alias{indikatorfigServer}
\title{Server-modul for indikatorfigurer i NNRR sin shiny-app på Rapporteket}
\usage{
indikatorfigServer(id, RegData, userRole, hvd_session)
}
\description{
Kun til bruk i Shiny
}
