% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagKodebok.R
\name{lagKodebok}
\alias{lagKodebok}
\title{Denne funksjonen tilrettelegger en kodebok for bruk i rapportering}
\usage{
lagKodebok()
}
\value{
En formatert kodebok
}
\description{
Denne funksjonen tilrettelegger en kodebok for bruk i rapportering
}
