% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrBeregnAndeler.R
\name{nnrrBeregnAndeler}
\alias{nnrrBeregnAndeler}
\title{Lag søylediagram som viser andeler av ulike variabler:}
\usage{
nnrrBeregnAndeler(
  RegData,
  valgtVar,
  datoFra = "2014-01-01",
  datoTil = "2050-12-31",
  enhetsUtvalg = 1,
  minald = 0,
  maxald = 120,
  erMann = 99,
  tverrfaglig = 99,
  minHSCL = 1,
  maxHSCL = 4,
  medikamenter = NULL,
  smerte = NULL,
  tolk = 99,
  valgtShus = "",
  reshID
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{enhetsUtvalg:}{0: Hele landet
1: Egen enhet mot resten av landet
2: Egen enhet}
}
\value{
En figur med beskrevet innhold
}
\description{
Denne funksjonen lager søylediagram som viser fordelingen til valgt variabel
}
