% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrBeregnAndelTid.R
\name{nnrrBeregnAndelTid}
\alias{nnrrBeregnAndelTid}
\title{Tidstrend av rate/andel for en gitt variabel}
\usage{
nnrrBeregnAndelTid(
  RegData,
  valgtVar = "tverrfaglig_behandlet",
  datoFra = "2014-01-01",
  datoTil = "2050-12-31",
  enhetsUtvalg = 1,
  datovar = "Besoksdato",
  minald = 0,
  maxald = 130,
  erMann = 99,
  outfile = "",
  reshID,
  tidsenhet = "Kvartal",
  maal = NA,
  maalnivaatxt = NA,
  tverrfaglig = 99,
  minHSCL = 1,
  maxHSCL = 4,
  medikamenter = NULL,
  smerte = NULL,
  tolk = 99,
  iArbeid = 99,
  valgtShus = "",
  ...
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{tidsenhet}{Tidsenhet for figur
'Aar' (Default)
'Kvartal'
'Mnd'}
}
\value{
En figur med tidsutvikling av andel over tid
}
\description{
Årlige, kvartalsmessige eller månedlige rater for valgt variabel.
Konfidensintervall kan inkluderes hvis ønskelig.
}
\details{
Konfidensintervallet er basert på Clopper Pearsons "eksakte" metode for binominalfordelt data.
}
