% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrBeregnGjsnPrePostGrVar.R
\name{nnrrBeregnGjsnPrePostGrVar}
\alias{nnrrBeregnGjsnPrePostGrVar}
\title{Beregn gjennomsnitt før og etter intervensjon fordelt på sykehus}
\usage{
nnrrBeregnGjsnPrePostGrVar(
  RegData,
  valgtVar,
  datoFra = "2000-01-01",
  datoTil = "2050-01-01",
  reshID,
  minald = 0,
  maxald = 120,
  erMann = 99,
  outfile = "",
  sammenlign = 1,
  enhetsUtvalg = 0,
  gr_var = "SykehusNavn"
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{gr_var}{Grupperingsvariabel}
}
\value{
plotparams En liste med verdier til generering av plot
}
\description{
Funksjon som beregner endring i en variabels gjennomsnitt før og etter intervensjonen
}
