% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrBeregnIndikator.R
\name{nnrrBeregnIndikator}
\alias{nnrrBeregnIndikator}
\title{Beregn indikatorer}
\usage{
nnrrBeregnIndikator(RegData, ind_id)
}
\arguments{
\item{RegData}{En dataramme med registerdata}
}
\description{
Inneholder logikken som definerer indikatorene til registeret
}
