% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrBrukerkontroller.R
\name{nnrrBrukerkontroller}
\alias{nnrrBrukerkontroller}
\title{Brukerkontroller for bruk i shinyapp}
\usage{
nnrrBrukerkontroller(RegData)
}
\arguments{
\item{RegData}{nnrr sin rapporteringsdata}
}
\description{
Definerer verdier til bruk i brukerkontroller for nnrr sin shinyapp.
}
