% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrFigAndelerPrePost.R
\name{nnrrFigAndelerPrePost}
\alias{nnrrFigAndelerPrePost}
\title{Søylediagram som fordeling av valgt variabel, målt både før og etter intervensjon}
\usage{
nnrrFigAndelerPrePost(RegData = 0, valgtVar, datoFra = "2000-01-01",
  datoTil = "2050-01-01", reshID, minald = 0, maxald = 120,
  erMann = 99, outfile = "", enhetsUtvalg = 1, preprosess = F,
  hentData = F)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}
}
\value{
Søylediagram som fordeling av valgt variabel, målt både før og etter intervensjonen
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels fordeling fra før til etter intervensjonen
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
