% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrFigAndelGrvar_6og12mnd.R
\name{nnrrFigAndelGrvar_6og12mnd}
\alias{nnrrFigAndelGrvar_6og12mnd}
\title{Andel for en gitt variabel gruppert etter grupperingsvariabel}
\usage{
nnrrFigAndelGrvar_6og12mnd(
  RegData,
  valgtVar1 = "opplevd_nytte_beh_6mnd",
  valgtVar2 = "opplevd_nytte_beh_12mnd",
  datoFra = "2014-01-01",
  datoTil = "2050-12-31",
  enhetsUtvalg = 1,
  datovar = "Besoksdato",
  minald = 0,
  maxald = 130,
  erMann = 99,
  outfile = "",
  reshID,
  inkl_konf = 0,
  grvar = "beh_spes",
  grtittel = "Behandling i spesialisthelsetjenesten"
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}
}
\value{
En figur med andel gruppert etter grupperingsvariabel
}
\description{
Første utkast, skal rafineres.
}
