% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrFigGjsnGrVarTid.R
\name{nnrrFigGjsnGrVarTid}
\alias{nnrrFigGjsnGrVarTid}
\title{Gjennomsnitt av valgt variabel pr grupperingsvariabel}
\usage{
nnrrFigGjsnGrVarTid(
  RegData,
  valgtVar = "PatientAge",
  tittel = "",
  datoFra = "2014-01-01",
  datoTil = "2050-12-31",
  minstekrav = NA,
  maal = NA,
  skriftStr = 1.3,
  pktStr = 1.4,
  minald = 0,
  maxald = 130,
  erMann = 99,
  outfile = "",
  gr_var = "SykehusNavn",
  terskel = 30,
  lavDG = "",
  decreasing = F,
  legPlass = "top",
  minstekravTxt = "Min.",
  maalTxt = "Mål",
  graaUt = NA,
  width = 800,
  height = 700,
  inkl_konf = F,
  maalretn = "hoy"
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}
}
\value{
En figur med gjennomsnitt av valgt variabel pr grupperingsvariabel
}
\description{
Denne funksjonen beregner gjennomsnitt for valgt variabel og plotter de per grupperingsvariabel (vanligvis sykehus) for
de siste trre år
}
\details{
Her kan detaljer skrives
}
