% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrFigGjsnPrePostGrVar.R
\name{nnrrFigGjsnPrePostGrVar}
\alias{nnrrFigGjsnPrePostGrVar}
\title{Søylediagram med gjennomsnitt før og etter intervensjon fordelt på sykehus}
\usage{
nnrrFigGjsnPrePostGrVar(RegData, valgtVar, datoFra = "2000-01-01",
  datoTil = "2050-01-01", reshID, minald = 0, maxald = 120, erMann = 99,
  outfile = "", enhetsUtvalg = 0, preprosess = F, hentData = F,
  gr_var = "SykehusNavn")
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{gr_var}{Velg grupperingsvariabel}
}
\value{
Søylediagram med gjennomsnitt før og etter intervensjon fordelt på sykehus
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels gjennomsnitt før og etter intervensjonen
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
