% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrHentRegData.R
\name{nnrrHentRegData}
\alias{nnrrHentRegData}
\title{Provide global dataframe for NNRR}
\usage{
nnrrHentRegData(datoFra = "2014-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}
}
\value{
RegData data frame
}
\description{
Provides NNRR data from staging
}

