% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrHentRegData.R
\name{nnrrHentTabell}
\alias{nnrrHentTabell}
\title{Fetch relevant table for NNRR}
\usage{
nnrrHentTabell(tabellnavn, datoFra, datoTil)
}
\arguments{
\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}
}
\value{
Data frame with relevant table
}
\description{
Provides NNRR data
}
