% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrPlotGjsnPrePostGrVar.R
\name{nnrrPlotGjsnPrePostGrVar}
\alias{nnrrPlotGjsnPrePostGrVar}
\title{Søylediagram med gjennomsnitt før og etter intervensjon fordelt på sykehus}
\usage{
nnrrPlotGjsnPrePostGrVar(
  plotparams,
  outfile = "",
  inkl_konf = 1,
  fargepalett = "BlaaOff",
  terskel = 5
)
}
\arguments{
\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{gr_var}{Velg grupperingsvariabel}
}
\value{
Søylediagram med gjennomsnitt før og etter intervensjon fordelt på sykehus
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels gjennomsnitt før og etter intervensjonen
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
