% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrPlotIndikator.R
\name{nnrrPlotIndikator}
\alias{nnrrPlotIndikator}
\title{Gi en visuell fremstilling av registerets indikatorer over tid}
\usage{
nnrrPlotIndikator(
  indikatordata,
  graaUt = NA,
  outfile = "",
  lavDG = NA,
  inkl_konf = F
)
}
\arguments{
\item{indikatordata}{En dataramme med følgende kolonner:
- ReshId
- year
- Teller
- Sykehusnavn}
}
\description{
Gi en visuell fremstilling av registerets indikatorer over tid
}
