% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrPreprosess.R
\name{nnrrPreprosess}
\alias{nnrrPreprosess}
\title{Preprosesser data for bruk i NNRR sine rapporter}
\usage{
nnrrPreprosess(RegData)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}
}
\value{
Et preprosessert datasett
}
\description{
Denne funksjonen gjor nodvendig preprosessering av NNRR sin data for bruk i rapporter
}
