% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrPrepVar.R
\name{nnrrPrepVar}
\alias{nnrrPrepVar}
\title{Preparer variabler for plotting}
\usage{
nnrrPrepVar(RegData, valgtVar)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabel som skal vises}
}
\value{
PrepData En liste med plotrelevante størrelser
}
\description{
Denne funksjonen grupperer og klargjør variabler for andelsplot
}
\details{
Her kan detaljer skrives
}

