% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrSoyleplot.R
\name{nnrrSoyleplot}
\alias{nnrrSoyleplot}
\title{Lag søylediagram som viser andeler av ulike variabler:}
\usage{
nnrrSoyleplot(plotdata, outfile = "", fargepalett = "BlaaRapp")
}
\arguments{
\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet}

\item{valgtVar}{- Variabel som skal vises}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{enhetsUtvalg:}{0: Hele landet
1: Egen enhet mot resten av landet
2: Egen enhet}
}
\value{
En figur med beskrevet innhold
}
\description{
Denne funksjonen lager søylediagram som viser fordelingen til valgt variabel
}
