% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrStabelGrVar.R
\name{nnrrStabelGrvar}
\alias{nnrrStabelGrvar}
\title{Plot stablede andeler per grupperingsvariabel}
\usage{
nnrrStabelGrvar(
  plotMatrise,
  grtxt = "",
  outfile = "",
  legendTxt = NA,
  tittel = "",
  xlab = " Antall pasienter",
  cexgr = 1,
  fargepalett = "BlaaOff",
  beside = F,
  revcol = F,
  revcol_legend = F
)
}
\description{
Stabelplot
}
