% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrTidsplot_6og12mnd.R
\name{nnrrTidsplot_6og12mnd}
\alias{nnrrTidsplot_6og12mnd}
\title{Tidstrend av rate/andel for en gitt variabel}
\usage{
nnrrTidsplot_6og12mnd(
  plotdata1,
  plotdata2,
  outfile = "",
  fargepalett = "BlaaRapp",
  inkl_konf = 1,
  maal = NA,
  maalnivaatxt = "Høy måloppnåelse",
  moderat = NA,
  moderatTxt = "Moderat måloppnåelse"
)
}
\arguments{
\item{outfile}{- navn på fil figuren skrives til, default '' dvs. skriv til skjerm}

\item{inkl_konf}{Inkluder konfidensintervall i figur
0: Nei
1: Ja
99: Bruk default som definert i nnrrPrepVar}

\item{tidsenhet}{Tidsenhet for figur
'Aar' (Default)
'Kvartal'
'Mnd'}
}
\value{
En figur med tidsutvikling av andel over tid
}
\description{
Årlige, kvartalsmessige eller månedlige rater for valgt variabel.
Konfidensintervall kan inkluderes hvis ønskelig.
}
\details{
Konfidensintervallet er basert på Clopper Pearsons "eksakte" metode for binominalfordelt data.
}
