% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrPlotIndikator.R
\name{plotlyIndikator}
\alias{plotlyIndikator}
\title{Gi en visuell fremstilling av registerets indikatorer over tid}
\usage{
plotlyIndikator(
  indikatordata,
  graaUt = NA,
  outfile = "",
  lavDG = NA,
  inkl_konf = F,
  ant_aar = 3
)
}
\arguments{
\item{indikatordata}{En dataramme med følgende kolonner:}
}
\description{
Gi en visuell fremstilling av registerets indikatorer over tid
}
