% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_samledok.R
\name{samledok_server}
\alias{samledok_server}
\title{Server-modul for Samledokumenter-fane i NNRR sin shiny-app på Rapporteket}
\usage{
samledok_server(id, reshID, RegData, userRole, hvd_session)
}
\value{
Modulfunksjoner til Samledokumenter
}
\description{
Kun til bruk i Shiny
}
