% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_startside.R
\name{startside}
\alias{startside}
\title{Serverdel av startside for NNRR app på Rapporteket}
\usage{
startside(input, output, session, usrRole)
}
\description{
Serverdel av startside for NNRR app på Rapporteket
}
