% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrHjelpefunksjoner.R
\name{strikkRnwAbo}
\alias{strikkRnwAbo}
\title{Generer kvartalsrapport og returner filnavn og sti til fil.}
\usage{
strikkRnwAbo(baseName, reshID = 0, datoTil = Sys.Date())
}
\description{
Generer kvartalsrapport og returner filnavn og sti til fil.
}
