% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sykehusvisning.R
\name{sykehusvisningServer}
\alias{sykehusvisningServer}
\title{Server-modul for sykehusvise figurer i NNRR sin shiny-app på Rapporteket}
\usage{
sykehusvisningServer(id, RegData, userRole, hvd_session)
}
\description{
Kun til bruk i Shiny
}
