% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_tidsvisning.R
\name{tidsvisningServer}
\alias{tidsvisningServer}
\title{Server-modul for tidsvisninger i NNRR sin shiny-app på Rapporteket}
\usage{
tidsvisningServer(id, reshID, RegData, userRole, hvd_session)
}
\description{
Kun til bruk i Shiny
}
