% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vask_diagnoser.R
\name{vask_diagnoser}
\alias{vask_diagnoser}
\title{Trekk ut icd-10 kode fra fritekst, og rydd i navn}
\usage{
vask_diagnoser(diagnoser_raa)
}
\value{
diagnoser_vasket En vektor av vaskede diagnoser
}
\description{
Denne funksjonen forsøker å trekke ut relevante icd-10 koder fra en vektor som også inneholder fritekst
}
\details{
Her kan detaljer skrives
}
