% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrIndikatorTabell.R
\name{nnrrIndikatorTabell}
\alias{nnrrIndikatorTabell}
\title{Fremstill tabell med indikator over 책r og fordelt p책 sykehus}
\usage{
nnrrIndikatorTabell(RegData, indikator = "Indikator")
}
\arguments{
\item{RegData}{Dataramme}

\item{indikator}{Navn p책 indikatorkolonne. Kolonne best책r av 0 og 1.}
}
\value{
Tabell
}
\description{
Tar som input en dataramme som minimum inneholder kolonnene Aar, SykehusNavn samt
en indikatorkolonne med navn man angir i funksjonskallet
}
