% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrUtvalg.R
\name{nnrrUtvalg}
\alias{nnrrUtvalg}
\title{Gjør utvalg av dataen}
\usage{
nnrrUtvalg(
  RegData,
  datoFra = "2000-01-01",
  datoTil = "2100-01-01",
  minald = 0,
  maxald = 120,
  erMann = 99,
  datovar = "Besoksdato",
  tverrfaglig = 99,
  minHSCL = 1,
  maxHSCL = 4,
  medikamenter = NA,
  smerte = NA,
  tolk = 99,
  iArbeid = 99,
  fargepalett = "BlaaRapp"
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}
}
\value{
Et filtrert datasett, samt en tekst som angir utvalget som er gjort
}
\description{
Denne funksjonen gjør utvalg av dataen basert på brukervalg
}
