% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrrUtvalg.R
\name{nnrrUtvalg}
\alias{nnrrUtvalg}
\title{Gjør utvalg av dataen}
\usage{
nnrrUtvalg(
  RegData,
  datoFra,
  datoTil,
  minald,
  maxald,
  erMann,
  datovar = "Besoksdato",
  fargepalett = "BlaaRapp"
)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{datoFra}{<- '2014-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2050-12-31'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, default 99, dvs. begge}
}
\value{
Et filtrert datasett, samt en tekst som angir utvalget som er gjort
}
\description{
Denne funksjonen gjør utvalg av dataen basert på brukervalg
}
